+++ test/unit/rmq_metadata/get_pdfminer_data.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python
 # Classification (U)
 
-"""Program:  get_pypdf2_data.py
+"""Program:  get_pdfminer_data.py
 
-    Description:  Unit testing of get_pypdf2_data in rmq_metadata.py.
+    Description:  Unit testing of get_pdfminer_data in rmq_metadata.py.
 
     Usage:
-        test/unit/rmq_metadata/get_pypdf2_data.py
+        test/unit/rmq_metadata/get_pdfminer_data.py
 
     Arguments:
 
@@ -43,7 +43,7 @@ class Logger(object):
     Methods:
         __init__ -> Initialize configuration environment.
         log_info -> log_info method.
-        log_warn -> log_warn method.
+        log_err -> log_err method.
 
     """
 
@@ -82,11 +82,11 @@ class Logger(object):
 
         self.data = data
 
-    def log_warn(self, data):
+    def log_err(self, data):
 
-        """Method:  log_warn
+        """Method:  log_err
 
-        Description:  log_warn method.
+        Description:  log_err method.
 
         Arguments:
             (input) data -> Log entry.
@@ -134,6 +134,7 @@ class CfgTest(object):
         self.stanford_jar = "DIRECTORY_PATH/stanford-ner.jar"
         self.encoding = "utf-8"
         self.token_types = ["LOCATION", "PERSON", "ORGANIZATION"]
+        self.textract_codes = ["utf-8", "ascii", "iso-8859-1"]
         self.queue_list = [
             {"queue": "rmq_metadata_unit_test",
              "routing_key": "ROUTING_KEY",
@@ -154,10 +155,11 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
-        test_extract_failed -> Test with categorized data returned.
-        test_extract_success -> Test with extraction successful.
-        test_no_categorized_data -> Test with no categorized data returned.
-        test_categorized_data -> Test with categorized data returned.
+        test_extract_failure -> Test with failure of extraction.
+        test_extract_success -> Test with successful extraction.
+        test_categorized_text2 -> Test with no categorized text.
+        test_categorized_text -> Test with categorized text.
+        test_get_pdfminer_data -> Test with extracting data.
 
     """
 
@@ -171,113 +173,136 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.f_name = "FileName.pdf"
-        self.rawtext = u'Intheseunprecedentedtimeswewanttomakesurewecankeepin'
-        self.tokens = [u'2.08', u'%', u'BalanceTransfer22.9', u'%', u'1.74']
-        self.categorized_text = [(u',', u'O'), (u'London', u'LOCATION'),
-                                 (u',', u'O'), (u'SW1W9AX', u'O')]
-
-        self.final_data = {
-            u'ORGANIZATION': [u'PAYPAL', u'HOMEBASE'],
-            u'LOCATION': [u'London', u'Brighton', u'England', u'Wales'],
-            u'PERSON': [u'John Street', u'SMITH'],
-            'filename': 'mail2rmq_mail2rmq_file_20200924082706.1493.pdf',
-            'datetime': '20200924_082717'}
         self.cfg = CfgTest()
+        self.f_name = "Filename.pdf"
+        self.text = "This is extracted text"
+        self.tokens = ["Tokens Strings", "Another token string"]
+        self.categorized_text = [("Token", "O"), ("Strings", "O")]
+        self.categorized_text2 = []
+        self.final_data = [("Place", "LOCATION")]
+        self.results = [("Place", "LOCATION")]
         self.logger = Logger("Name", "Name", "INFO", "%(asctime)s%(message)s",
                              "%m-%d-%YT%H:%M:%SZ|")
 
     @mock.patch("rmq_metadata.summarize_data")
     @mock.patch("rmq_metadata.find_tokens")
     @mock.patch("rmq_metadata.word_tokenize")
-    @mock.patch("rmq_metadata.read_pdf")
-    def test_extract_failed(self, mock_read, mock_token, mock_find, mock_summ):
+    @mock.patch("rmq_metadata.pdf_to_string")
+    def test_extract_failure(self, mock_extract, mock_token, mock_find,
+                              mock_sum):
 
-        """Function:  test_extract_failed
+        """Function:  test_extract_failure
 
-        Description:  Test with categorized data returned.
+        Description:  Test with failure of extraction.
 
         Arguments:
 
         """
 
-        mock_read.return_value = (False, "")
+        mock_extract.return_value = (False, "")
         mock_token.return_value = self.tokens
         mock_find.return_value = self.categorized_text
-        mock_summ.return_value = self.final_data
+        mock_sum.return_value = self.final_data
 
         self.assertEqual(
-            rmq_metadata.get_pypdf2_data(self.f_name, self.cfg, self.logger),
+            rmq_metadata.get_pdfminer_data(self.f_name, self.cfg, self.logger),
             (False, []))
 
     @mock.patch("rmq_metadata.summarize_data")
     @mock.patch("rmq_metadata.find_tokens")
     @mock.patch("rmq_metadata.word_tokenize")
-    @mock.patch("rmq_metadata.read_pdf")
-    def test_extract_success(self, mock_read, mock_token, mock_find,
-                             mock_summ):
+    @mock.patch("rmq_metadata.pdf_to_string")
+    def test_extract_success(self, mock_extract, mock_token, mock_find,
+                             mock_sum):
 
         """Function:  test_extract_success
 
-        Description:  Test with extraction successful.
+        Description:  Test with successful extraction.
 
         Arguments:
 
         """
 
-        mock_read.return_value = (True, self.rawtext)
+        mock_extract.return_value = (True, self.text)
         mock_token.return_value = self.tokens
         mock_find.return_value = self.categorized_text
-        mock_summ.return_value = self.final_data
+        mock_sum.return_value = self.final_data
 
         self.assertEqual(
-            rmq_metadata.get_pypdf2_data(self.f_name, self.cfg, self.logger),
-            (True, self.final_data))
+            rmq_metadata.get_pdfminer_data(self.f_name, self.cfg, self.logger),
+            (True, self.results))
 
+    @mock.patch("rmq_metadata.summarize_data")
     @mock.patch("rmq_metadata.find_tokens")
     @mock.patch("rmq_metadata.word_tokenize")
-    @mock.patch("rmq_metadata.read_pdf")
-    def test_no_categorized_data(self, mock_read, mock_token, mock_find):
+    @mock.patch("rmq_metadata.pdf_to_string")
+    def test_categorized_text2(self, mock_extract, mock_token, mock_find,
+                               mock_sum):
 
-        """Function:  test_no_categorized_data
+        """Function:  test_categorized_text2
 
-        Description:  Test with no categorized data returned.
+        Description:  Test with no categorized text.
 
         Arguments:
 
         """
 
-        mock_read.return_value = (True, self.rawtext)
+        mock_extract.return_value = (True, self.text)
         mock_token.return_value = self.tokens
-        mock_find.return_value = []
+        mock_find.return_value = self.categorized_text2
+        mock_sum.return_value = self.final_data
 
         self.assertEqual(
-            rmq_metadata.get_pypdf2_data(self.f_name, self.cfg, self.logger),
+            rmq_metadata.get_pdfminer_data(self.f_name, self.cfg, self.logger),
             (True, []))
 
     @mock.patch("rmq_metadata.summarize_data")
     @mock.patch("rmq_metadata.find_tokens")
     @mock.patch("rmq_metadata.word_tokenize")
-    @mock.patch("rmq_metadata.read_pdf")
-    def test_categorized_data(self, mock_read, mock_token, mock_find,
-                              mock_summ):
+    @mock.patch("rmq_metadata.pdf_to_string")
+    def test_categorized_text(self, mock_extract, mock_token, mock_find,
+                               mock_sum):
 
-        """Function:  test_categorized_data
+        """Function:  test_categorized_text
 
-        Description:  Test with categorized data returned.
+        Description:  Test with categorized text.
 
         Arguments:
 
         """
 
-        mock_read.return_value = (True, self.rawtext)
+        mock_extract.return_value = (True, self.text)
         mock_token.return_value = self.tokens
         mock_find.return_value = self.categorized_text
-        mock_summ.return_value = self.final_data
+        mock_sum.return_value = self.final_data
 
         self.assertEqual(
-            rmq_metadata.get_pypdf2_data(self.f_name, self.cfg, self.logger),
-            (True, self.final_data))
+            rmq_metadata.get_pdfminer_data(self.f_name, self.cfg, self.logger),
+            (True, self.results))
+
+    @mock.patch("rmq_metadata.summarize_data")
+    @mock.patch("rmq_metadata.find_tokens")
+    @mock.patch("rmq_metadata.word_tokenize")
+    @mock.patch("rmq_metadata.pdf_to_string")
+    def test_get_pdfminer_data(self, mock_extract, mock_token, mock_find,
+                               mock_sum):
+
+        """Function:  test_get_pdfminer_data
+
+        Description:  Test with extracting data.
+
+        Arguments:
+
+        """
+
+        mock_extract.return_value = (True, self.text)
+        mock_token.return_value = self.tokens
+        mock_find.return_value = self.categorized_text
+        mock_sum.return_value = self.final_data
+
+        self.assertEqual(
+            rmq_metadata.get_pdfminer_data(self.f_name, self.cfg, self.logger),
+            (True, self.results))
 
 
 if __name__ == "__main__":