+++ test/unit/rmq_metadata/unit_test_run.sh
@@ -13,6 +13,7 @@ test/unit/rmq_metadata/_validate_files.py
 test/unit/rmq_metadata/create_metadata.py
 test/unit/rmq_metadata/extract_pdf.py
 test/unit/rmq_metadata/find_tokens.py
+test/unit/rmq_metadata/get_pdfminer_data.py
 test/unit/rmq_metadata/get_pypdf2_data.py
 test/unit/rmq_metadata/get_textract_data.py
 test/unit/rmq_metadata/help_message.py
@@ -20,6 +21,7 @@ test/unit/rmq_metadata/main.py
 test/unit/rmq_metadata/merge_data.py
 test/unit/rmq_metadata/monitor_queue.py
 test/unit/rmq_metadata/non_proc_msg.py
+test/unit/rmq_metadata/pdf_to_string.py
 test/unit/rmq_metadata/process_msg.py
 test/unit/rmq_metadata/read_pdf.py
 test/unit/rmq_metadata/run_program.py